requirements.txt
-i https://pypi.org/simple
alembic==1.4.3
awscli==1.18.166
boto3==1.16.6
botocore==1.19.6
click==7.1.2
colorama==0.4.3; python_version != '3.4'
docutils==0.15.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
faker==4.14.0
flask-cors==3.0.8
flask-jwt-extended==3.24.1
flask-login==0.5.0
flask-migrate==2.5.3
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
jinja2==2.11.2
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
mako==1.1.3
markupsafe==1.1.1
pyasn1==0.4.8
pyjwt==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pyyaml==5.3.1; python_version != '3.4'
rsa==4.5; python_version != '3.4'
s3transfer==0.3.3
six==1.15.0
sqlalchemy==1.3.19
text-unidecode==1.3
urllib3==1.25.11; python_version != '3.4'
werkzeug==1.0.1
wtforms==2.3.3
